% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{observe_random_vector}
\alias{observe_random_vector}
\title{Simulate observing a random vector multiple times}
\usage{
observe_random_vector(n = 1, dim = 1, distribution_sampler = stats::rnorm)
}
\arguments{
\item{n}{A positive integer giving the number of times to observe the
vector.}

\item{dim}{A positive integer giving the dimension of the random vector.}

\item{distribution_sampler}{A function of \code{n} that samples from the
distribution that each component of the vector is assumed to have.}
}
\value{
An \code{n} by \code{dim} matrix whose rows are the observations of the
simulated random vector.
}
\description{
This function simulates observing a \code{dim}-dimensional random vector
\code{n} times, whose components are independent, each of which can be
sampled from with the function \code{distribution} (a function of \code{n}).
}
