% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_attentions}
\alias{plot_attentions}
\title{Make a density plot of rescaled vector attention scores}
\usage{
plot_attentions(
  attention_matrices,
  legend_texts = 1:length(attention_matrices),
  scalar_attention_function = `\%*\%`,
  key_index = 1
)
}
\arguments{
\item{attention_matrices}{An attention matrix or a list of attention
matrices, each of which will ordinarily be the output of
\verb{simulate attentions}. That is, each attention matrix should have its
(i,j)-th entry be the scalar attention of the i-th query on the j-th key.}

\item{legend_texts}{A list (recycled if need be) of labels, one for each
attention matrix in \code{attention_matrices}. These labels will appear in the
legend of the plot to indicate which density plot is which.}

\item{scalar_attention_function}{A function of two variables (query and
key) that computes the scalar attention of the query on the key.}

\item{key_index}{The column numbers in the attention matrices that should
be included in the kernel density estimate plot.}
}
\value{
A \code{ggplot2} kernel density estimate plot of attention scores.
}
\description{
This function makes a kernel density estimate plot of the rescaled vector
attention scores for all the given queries and one or more keys (placed
into the rows of a matrix). By default only the first component of the
rescaled vector attention scores is plotted, but you can plot specific
key numbers by putting those numbers into the \code{key_index} vector. You can
also plot all components by setting \code{key_index} to \code{NULL}.
}
