% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scaled_softmax}
\alias{scaled_softmax}
\title{Compute softmax of a vector after dividing each component by something.}
\usage{
scaled_softmax(x, divide_by = 1)
}
\arguments{
\item{x}{A vector whose scaled softmax is to be computed.}

\item{divide_by}{A quantity to divide x by before applying softmax. This
is usually a scalar, but can in theory be a vector.}
}
\value{
A vector containing the softmax of \code{x} divided by \code{divide_by}.
}
\description{
This function takes a vector, divides all of its arguments by
\code{divided_by} (which defaults to 1), and then returns the softmax of the
resulting vector.
}
