% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simulate_attentions}
\alias{simulate_attentions}
\title{Simulate queries and keys and compute their attention scores}
\usage{
simulate_attentions(
  distribution_sampler = stats::rnorm,
  n_keys = 1,
  n_queries = 1,
  key_dimension = 1,
  rescaling_function = function(keys, x) x,
  scalar_attention_function = `\%*\%`
)
}
\arguments{
\item{distribution_sampler}{A function of the sample size \code{n} that
samples from the distribution that each component is to have.}

\item{n_keys}{A positive integer specifying the number of keys to simulate}

\item{n_queries}{A positive integer specifying the number of queries to
simulate.}

\item{key_dimension}{A positive integer specifying the number of
components each key and query is to have.}

\item{rescaling_function}{A function whose arguments are a matrix whose
rows are keys (\code{keys}) and a vector whose dimension is the number of
keys (\code{x}). This function is
applied to the keys and to the vector of attention scores
when computing the rescaled attention vector. For each set of keys, it is
a function from an n_key-dimensional space itself, although for some
types of simulations the codomain should be the standard
(n_key-1)-dimensional simplex in that space.}

\item{scalar_attention_function}{A function of two vectors (each of
dimension \code{key_dimension}) that returns a real number.}
}
\value{
An \code{n_queries} by \code{n_keys} matrix whose (i,j) entry is the
scalar attention of the i-th query on the j-th key.
}
\description{
This function computes a matrices of attention scores for simulated queries
and keys, one matrix for each function in the \code{distribution_samplers} list.
}
